requirements.txt
streamlit
numpy
pandas
plotly
matplotlib
Pillow